[W5a] bc-io: fuse file_*.h + add walk_sequential + opacify POSIX types
Split tactique de W5 (P5.3 dirent_reader opaque traite en W5b separee
pour limiter le risque cross-tools).

P5.1 — fusion file_*.h
  Avant : 3 headers fragmentes (bc_io_file.h umbrella + bc_io_file_path.h
  + bc_io_file_open.h).
  Apres : 1 header bc_io_file.h consolide (61 LOC).
  bc_io_file_inode.h reste separe (struct opaque dediee).
  Suppression de 2 headers publics : bc_io_file_path.h, bc_io_file_open.h.
  meson.build install_headers mis a jour.
  10 -> 8 headers publics.

P5.2 — bc_io_walk_sequential expose
  Nouveau bool bc_io_walk_sequential(const config*, stats*) dans walk.h.
  Impl dans src/walk/bc_io_walk_sequential.c (218 LOC) reutilise
  bc_io_dirent_reader + bc_io_file_*. Ignorent concurrency_context.
  7 tests cmocka couvrant config invalide / vide / flat / nested /
  filter / root inexistant / visit failure abort.

P5.4 — opacification POSIX types dans walk_entry
  bc_io_walk_entry_t.device_id / inode_number passent de dev_t/ino_t
  a unsigned long long. Impl walk_parallel + walk_sequential cast
  les valeurs raw. <sys/types.h> retire de bc_io_walk.h public.
  Empeche fuite des types POSIX dans l'API publique.

Pipeline 27 gates verts :
- format / build/test release+debug.
- asan / tsan / ubsan / memcheck PASS.
- cppcheck PASS.

Bench delta vs baseline (median 3 runs) :
- bc_io_walk_parallel files=10000 : ratio bc/nftw 2.83x baseline ->
  2.88-3.04x mesure (variance dans bruit machine, run 3 = 34.76 ms
  vs baseline 35.45 ms, gain).
- inode_set insert (100k) : 59 -> 57 ns/op (-3%, gain).
- inode_set contains (1M) : 23 -> 22 ns/op (-4%, gain).
- bench_stream_read buf=4096 : 14.5 -> 13.3 ms (gain).
Aucune regression > 2% sur les ratios.

Cascade DAG :
- libs downstream (bc-runtime, bc-hrbl) : build + test + reinstall PASS.
- tools cascade : bc-hash (3 includes consolides) + bc-duplicate
  (2 includes consolides + 3 casts dev_t/ino_t) — PRs soeurs synchros.
- bc-seek / bc-integrity / bc-count : cascade rebuild PASS sans
  modification (comparaisons egalite ne declenchent pas -Wconversion
  sur narrowing dev_t/ino_t -> unsigned long long).
- bc-text : skip (autre conversation user).

Audit doctrine :
- 0 string.h primitive ajoutee.
- 0 macro template-like.
- 0 module sans consumer.

P5.3 (dirent_reader opaque) reste a faire en W5b separee — 4 tools
necessitent migration stack-alloc bc_io_dirent_reader_t reader vers
heap-alloc create/destroy.

Refs: audits/2026-04-30-code-review-bc-libraries.md sect. C.2/C.3/E.1/E.5/E.6
      audits/2026-04-30-plan-implementation-bc-libraries.md sect. W5

+++ src/walk/bc_io_walk_parallel.c
@@ -4,7 +4,6 @@
 
 #include "bc_io_dirent_reader.h"
 #include "bc_io_file.h"
-#include "bc_io_file_path.h"
 
 #include "bc_concurrency.h"
 #include "bc_concurrency_signal.h"
@@ -184,8 +183,8 @@ static void bc_io_walk_process_directory(bc_io_walk_shared_t* shared, const char
             .kind = entry_kind,
             .file_size = resolved_file_size,
             .depth = directory_depth + 1U,
-            .device_id = resolved_device_id,
-            .inode_number = resolved_inode_number,
+            .device_id = (unsigned long long)resolved_device_id,
+            .inode_number = (unsigned long long)resolved_inode_number,
             .modification_time_seconds = resolved_modification_time,
             .permission_mask = resolved_permission_mask,
             .stat_populated = stat_populated,
